# schemas/api/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FutureDiff Local API",
    "version": "1.1",
    "description": "Private Unix-socket HTTP API. Agent-safe authority is declared per operation."
  },
  "paths": {
...
    "/v1/transactions/{id}/abort": {
      "post": {
        "operationId": "transaction_abort",
        "summary": "transaction abort",
        "x-futurediff-agent-safe": false,
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "4XX": {
            "description": "Request rejected"
          },
          "5XX": {
            "description": "Server or policy failure"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        }
      }
    },
    "/v1/transactions/{id}/access": {
      "get": {
        "operationId": "transaction_access_list",
        "summary": "transaction access list",
        "x-futurediff-agent-safe": false,
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "4XX": {
            "description": "Request rejected"
          },
          "5XX": {
            "description": "Server or policy failure"
          }
        }
      }
    },
    "/v1/transactions/{id}/access/{principalID}": {
      "delete": {
        "operationId": "transaction_access_revoke",
        "summary": "transaction access revoke",
        "x-futurediff-agent-safe": false,
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "4XX": {
            "description": "Request rejected"
          },
          "5XX": {
            "description": "Server or policy failure"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        }
      },
      "put": {
        "operationId": "transaction_access_grant",
        "summary": "transaction access grant",
        "x-futurediff-agent-safe": false,
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "4XX": {
            "description": "Request rejected"
          },
          "5XX": {
            "description": "Server or policy failure"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        }
      }
    },
    "/v1/transactions/{id}/approval-material": {
      "get": {
        "operationId": "approval_material",
        "summary": "approval material",
        "x-futurediff-agent-safe": false,
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "4XX": {
            "description": "Request rejected"
          },
          "5XX": {
            "description": "Server or policy failure"
          }
        }
      }
    },
    "/v1/transactions/{id}/approve": {
      "post": {
        "operationId": "transaction_approve",
        "summary": "transaction approve",
        "x-futurediff-agent-safe": false,
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "4XX": {
            "description": "Request rejected"
          },
          "5XX": {
            "description": "Server or policy failure"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        }
      }
    },
    "/v1/transactions/{id}/commit": {
      "post": {
        "operationId": "transaction_commit",
        "summary": "transaction commit",
        "x-futurediff-agent-safe": false,
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "4XX": {
            "description": "Request rejected"
          },
          "5XX": {
            "description": "Server or policy failure"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        }
      }
    },
...
    "/v1/transactions/{id}/recover": {
      "post": {
        "operationId": "transaction_recover",
        "summary": "transaction recover",
        "x-futurediff-agent-safe": false,
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "4XX": {
            "description": "Request rejected"
          },
          "5XX": {
            "description": "Server or policy failure"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        }
      }
    },
...
  },
...
}